+++ CHANGELOG.md
@@ -23,7 +23,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
--
+- Fixed issue where Rectangle line renderer did not respect z order
+
 ### Updates
 
 -